refactor: move Noaa config models to reformatters.noaa.models and add accumulation fields
